fico-backend/requirements.txt
fastapi
uvicorn
pytest
httpx
google-auth
email-validator
requests
